fix: 仪表板 title 修改相关bug修复
