+++ frontend/src/views/chart/components/normal/LabelNormal.vue
@@ -11,7 +11,7 @@
       style="cursor: default;display: block;"
     >
       <div>
-        <p style="padding:6px 4px 0;margin: 0;overflow: hidden;white-space: pre;text-overflow: ellipsis;display: inline;">{{ chart.title }}</p>
+        <chart-title-update :chart-info="chartInfo" />
         <title-remark
           v-if="chart.render && chart.render === 'antv' && remarkCfg.show"
           style="text-shadow: none!important;"
@@ -47,10 +47,11 @@ import eventBus from '@/components/canvas/utils/eventBus'
 import { formatterItem, valueFormatter } from '@/views/chart/chart/formatter'
 import TitleRemark from '@/views/chart/view/TitleRemark'
 import { DEFAULT_SIZE, DEFAULT_TITLE_STYLE } from '@/views/chart/chart/chart'
+import ChartTitleUpdate from '../ChartTitleUpdate.vue'
 
 export default {
   name: 'LabelNormal',
-  components: { TitleRemark },
+  components: { TitleRemark, ChartTitleUpdate },
   props: {
     chart: {
       type: Object,
@@ -110,12 +111,10 @@ export default {
     }
   },
   computed: {
-    // bg_class() {
-    //   return {
-    //     background: hexColorToRGBA('#ffffff', 0),
-    //     borderRadius: this.borderRadius
-    //   }
-    // }
+    chartInfo() {
+      const { id, title } = this.chart
+      return { id, title }
+    }
   },
   watch: {
     chart() {
